Add support for extensions and imports
Uses new syntax for extension imports, as these must be distinguished
from regular Scala imports. An extension does not create a Scala type,
so a regular Scala import would complain.

At the current state only types can be imported, which are in a file of
the same name (as in Java), as the Sugarj compiler expects this.

Pretty printing table needs to be overhauled to produce correct Scala
syntax. This commit only tweaks imports to not print a newline after the
import keyword, as this is not correct Scala syntax. The list of
statements in a template body are not separated by newlines, which
requires the use of ";" after each statement.